refactor(project-services): type compose config args contract
