+++ tests/atelier/services/project/test_services.py
@@ -35,7 +35,7 @@ def test_compose_and_provider_contract_failures() -> None:
         enlistment_path: str,
         origin: str | None,
         origin_raw: str | None,
-        args: object | None,
+        args: InitProjectArgs | None,
         *,
         prompt_missing_only: bool = False,
         raw_existing: dict | None = None,
@@ -111,7 +111,7 @@ def test_compose_and_provider_failures_capture_system_exit_exception() -> None:
         enlistment_path: str,
         origin: str | None,
         origin_raw: str | None,
-        args: object | None,
+        args: InitProjectArgs | None,
         *,
         prompt_missing_only: bool = False,
         raw_existing: dict | None = None,
